updated datastream to have right prime database and fixed some main stuff/uncommented

+++ PseudoAPI_Datastream.py
@@ -159,15 +159,32 @@ def getcurrentmindata(connection, pricesymbols):
     add_row(connection, 'volumes', volumes, priceSymbols)
 
 #gives the databases 2 hours of data for each datatype
-def primeDatabase(connections, priceSymbols):
+def primeDatabase(connections, priceSymbols, params):
     """
     :param connections:
     :param priceSymbols: the price symbols to use
     :return:
     """
 
-    global buffertimestart
+    #global buffertimestart
 
+    #one day in ms
+    ONE_DAY = 86400000
+
+    #one day in min
+    ONE_DAY_MIN = 1440
+
+    #one third day in ms
+    ONE_THIRD_DAY = 28800000
+
+    #one third day in min
+    ONE_THIRD_MIN = 480
+
+    #one minute in ms
+    ONE_MIN_MS = 60000
+
+    #one second in ms
+    ONE_SEC_MS = 1000
 
     #grabbing the starttime of the data desired and the current time (endtime)
     endTime = requests.get("https://api.binance.com/api/v1/time")
@@ -182,19 +199,25 @@ def primeDatabase(connections, priceSymbols):
     lowpricedict = {}
     volumedict = {}
 
-    for x in range(ONE_DAY_MIN):
-        x = ONE_THIRD_MIN
-       
-        # set up the dicts to be made of lists where the index is the minute associated with the list of values
-        # and the values of each list correspond to the currencies in order from price symbols
-        # these are made this way to facilitate easy transfer to the tables of the database
-        for minute in range(x):
-            openpricedict.update({minute: []})
-            closepricedict.update({minute: []})
-            highpricedict.update({minute: []})
-            lowpricedict.update({minute: []})
-            volumedict.update({minute: []})
+    x = ONE_THIRD_MIN
+    
+    buffertimestart = time.time()
 
+    # set up the dicts to be made of lists where the index is the minute associated with the list of values
+    # and the values of each list correspond to the currencies in order from price symbols
+    # these are made this way to facilitate easy transfer to the tables of the database
+    for minute in range(params['minutestoprime']):
+        openpricedict.update({minute: []})
+        closepricedict.update({minute: []})
+        highpricedict.update({minute: []})
+        lowpricedict.update({minute: []})
+        volumedict.update({minute: []}) 
+
+    while(x <= params['minutestoprime']):
+        minute = x - ONE_THIRD_MIN
+
+        print('Minute: {}, X: {}'.format(minute, x))
+       
         #iterate through the dictionary of price symbols and store the five kinds of data in their corresponding dictionaries
         for currencyname in priceSymbols:
             #store 2 hours of data for the five categories to prime the database
@@ -202,32 +225,85 @@ def primeDatabase(connections, priceSymbols):
             data = requests.get("https://api.binance.com/api/v1/klines", params=parameters)
             data = data.json()
 
-            #iterate through the 2 hours of data and store it in ascending order (oldest to newest)
-            min = 0
+            #iterate through the data and store it in ascending order (oldest to newest)
+            minute = x - ONE_THIRD_MIN
             for interval in data:
-                openpricedict[min].append(interval[1])
-                closepricedict[min].append(interval[4])
-                highpricedict[min].append(interval[2])
-                lowpricedict[min].append(interval[3])
-                volumedict[min].append(interval[5])
+                openpricedict[minute].append(interval[1])
+                closepricedict[minute].append(interval[4])
+                highpricedict[minute].append(interval[2])
+                lowpricedict[minute].append(interval[3])
+                volumedict[minute].append(interval[5])
 
-                min+=1
+                minute += 1
 
         x += ONE_THIRD_MIN
         endTime = startTime
         startTime -= ONE_THIRD_DAY
-    '''
+    
     #grabbing the time after the last set of data is stored
-    buffertimestart = time.time()
+    buffertimeend = time.time()
+    
+    #calculate the time difference in seconds between when we started and ended priming the database
+    secondsTaken = buffertimeend - buffertimestart
 
+    #convert these two to milliseconds so they can be used by binance
+    buffertimestart = int(buffertimestart) * 1000
+    buffertimeend = int(buffertimeend) * 1000        
+
+
+    for currencyname in priceSymbols:
+        #store the overflow minute data for the five categories to prime the database
+        parameters = {'symbol': currencyname, 'startTime': buffertimestart, 'endTime': buffertimeend, 'interval': '1m'}
+        data = requests.get("https://api.binance.com/api/v1/klines", params=parameters)
+        data = data.json()
+
+        #how many minutes binance says has passed
+        binanceMin = len(data)
+
+        #this program takes between 2 and 3 minutes sometimes binance returns 3 so if they returned 3 we need to wait to reflect 3 minutes passing 
+        if(len(data) == 3):
+            delta = 180 - secondsTaken
+            time.sleep(delta)
+            secondsTaken = 180
+
+        #add the minutes to prime so that we can add to the end of the dictionary
+        binanceMin += params['minutestoprime']
+
+        #if the new keys haven't been created create them
+        if(params['minutestoprime'] not in openpricedict):
+            for minute in range(params['minutestoprime'], binanceMin):
+                openpricedict[minute] = []
+                closepricedict[minute] = []
+                highpricedict[minute] = []
+                lowpricedict[minute] = []
+                volumedict[minute] = []
+
+        print('Length returned: {}'.format(len(data)))
+        x = 0
+        #iterate through and append the actual data to the dictionaries
+        for minute in range(params['minutestoprime'], binanceMin):
+            openpricedict[minute].append(data[x][1])
+            closepricedict[minute].append(data[x][4])
+            highpricedict[minute].append(data[x][2])
+            lowpricedict[minute].append(data[x][3])
+            volumedict[minute].append(data[x][5])
+
+            x += 1
+    
+    print('exited loop')
+    logging.info('Open Price Dict: {}'.format(str(openpricedict)))
+
+    #print('Open Price Dict: {}'.format(openpricedict))
     #add each row of data to the five tables of the database
-    for rownum in range(TWO_HOURS_MIN):
-        #storre the list of values for the current row (minute) in the format used to create a new table row
+    for rownum in range(params['minutestoprime']): 
+       #storre the list of values for the current row (minute) in the format used to create a new table row
         opens = (openpricedict[rownum]);
         closes = (closepricedict[rownum]);
         highs = (highpricedict[rownum]);
         lows = (lowpricedict[rownum]);
         volumes = (volumedict[rownum]);
+
+        #print('Open Price Dict: {}'.format(opens))
 
         #pass the new lists of values to the functions that append them as new rows to each database
         add_row(connections, 'openprices', opens, priceSymbols)
@@ -235,7 +311,9 @@ def primeDatabase(connections, priceSymbols):
         add_row(connections, 'highprices', highs, priceSymbols)
         add_row(connections, 'lowprices', lows, priceSymbols)
         add_row(connections, 'volumes', volumes, priceSymbols)
-    '''
+    
+    connections.commit()
+
 #creates a connection with the specified database file
 def create_connection_db(db_file):
     """
@@ -572,13 +650,12 @@ def main():
 
     cursor = connection.cursor()
 
-    #uncomment if you want to clear the database before using
-    """
-    for name in tablenames:
-        delete_rows(connection, name)
-        
-    connection.commit()
-    """
+    if(params['freshrun']):
+        for name in tablenames:
+            delete_rows(connection, name)
+            
+        connection.commit()
+    
 
     numRows = getNumRows(cursor, 'openprices')
 
@@ -589,10 +666,10 @@ def main():
     #uncommon if you want to prime the database (TODO use to prime with params['hourprime'] = hourstoprime and params['freshrun'] = true
 
     #set the database up with 240 minutes of data
-    primeDatabase(connection, priceSymbols)
+    primeDatabase(connection, priceSymbols, params)
 
     #set the mins passed to reflect the new data
-    params['mins']+=240
+    params['mins'] += params['minutestoprime']
